# ulster-delt-investment/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Ulster Delt Investment API
  version: 1.0.0
  description: API documentation for Ulster Delt Investment banking platform
servers:
  - url: http://localhost:3000
    description: Development server
components:
  schemas:
...
    Account:
      type: object
      properties:
        _id:
          type: string
          description: The account's unique identifier
        userId:
          type: string
          description: The ID of the user who owns this account
        accountNumber:
          type: string
          description: The account number
        balance:
          type: number
          description: The current balance
        currency:
          type: string
          description: The account currency
        type:
          type: string
          enum: [savings, checking, investment]
          description: The type of account
        status:
          type: string
          enum: [active, inactive, frozen]
          description: The account status
        createdAt:
          type: string
          format: date-time
          description: The date the account was created
        updatedAt:
          type: string
          format: date-time
          description: The date the account was last updated
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/accounts:
    get:
      tags:
        - Accounts
      summary: Get user's accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized

    post:
      tags:
        - Accounts
      summary: Create a new account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - currency
              properties:
                type:
                  type: string
                  enum: [savings, checking, investment]
                currency:
                  type: string
      responses:
        '201':
          description: Account created successfully
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
  /api/accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get account details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
        '404':
          description: Account not found

    put:
      tags:
        - Accounts
      summary: Update account
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, inactive, frozen]
      responses:
        '200':
          description: Account updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Account not found
...